requirements.txt
# Core dependencies for Protein Analyzer
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
biopython==1.83